% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artif_spo.R
\name{artif_spo}
\alias{artif_spo}
\title{Artificial Event Origins}
\usage{
artif_spo(poly, n_origin=50, resistance_feat = NULL,
n_foci=5, foci_separation = 10, p_ratio)
}
\arguments{
\item{poly}{(An sf or S4 object)
A polygon shapefile within which
event origins are to be situated.}

\item{n_origin}{(an integer) Value specifying
the number of event origins to synthetize.
Default:\code{50}. Value specified here has the greatest
influence on the computational time.}

\item{resistance_feat}{(An S4 object) Optional
shapefile representing spaces across landscape
within which event
origins are not allowed. Default: \code{NULL}.}

\item{n_foci}{(an integer) Value indicating the number of
focal points amongst event origins. \code{n_foci} will usually be
smaller than \code{n_origin}.}

\item{foci_separation}{(an integer) A percentage value indicating
indicating the nearness of focal points from one another. A \code{0}
separation indicates that focal points are in close proximity
of one another, while a \code{100} indicates focal points being
evenly distributed across space.}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio.
For example, a value of \code{20}
implies a \code{20:80} Pareto ratio.}
}
\value{
Returns event origins with their respective
strength (probability) values.
}
\description{
Simulate points to serve as event
origins across space. If provided, resistance features
are to be avoided. Each final
event origin point is assigned a probability value
indicating the strength of the origin.
}
\details{
Details of events origins:
{x,y locations, categories (i.e. focal and non-focal (others)
origins), and the probability values.
and non-focal point}.
The focal origins (\code{n_foci}) serve as the more dominant
origins (e.g. city centres), while the non-focal origins
(i.e. non-dominant) origin. The \code{foci_separation} indicates
how close (to one another) the dominant origins are.
If \code{resistance_feat} is provided, the features help
to prevent event origins from being
situated in the same locations occupied by the features.
}
\examples{
data(camden_boundary)
data(camden_landuse)
spo <- artif_spo(poly = camden_boundary, n_origin = 50,
resistance_feat = camden_landuse, n_foci=5,
foci_separation = 0, p_ratio=20)
}
