% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camden_boundary}
\alias{camden_boundary}
\title{A boundary shapefile}
\format{
A boundary file (ESRI format)
\itemize{
\item x: x coordinate
\item y: y coordinate
}
}
\usage{
camden_boundary
}
\description{
A boundary shapefile
}
\keyword{datasets}
