% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camden_crimes}
\alias{camden_crimes}
\title{Records of crimes of Camden Borough of London,
UK, 2021 (Source: https://data.police.uk/data/)}
\format{
A matrix containing three variables
\itemize{
\item x: x coordinate
\item y: y coordinate
\item date: date of occurence
\item type: type of crime
}
}
\usage{
camden_crimes
}
\description{
'Theft' and 'Criminal Damage' records
(Source: \code{police.uk})
of Camden Borough of London, UK
for the year 2021.
Note: Police.uk data is only accurate to monthly
scale (\code{yyyy-mm}). But, the data provided here has been
manipulated to include fake 'days of occurrence'.
So, caution should
be taken when interpreting the results based on
full date stamp (\code{yyyy-mm-dd}).
}
\keyword{datasets}
