% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camden_landuse}
\alias{camden_landuse}
\title{A land use shapefile of Camden
Borough of London, United Kingdom}
\format{
A boundary file (ESRI format)
\itemize{
\item value: Values indicating
the level of resistance of each
feature. Ranges from 0 (low resistance)
to 1 (high resistance).
\item type: Landuse type
}
}
\usage{
camden_landuse
}
\description{
A land use shapefile of Camden
Borough of London, United Kingdom
}
\keyword{datasets}
