% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camden_theft}
\alias{camden_theft}
\title{'theft-from-person' crime of Camden Borough, London, UK}
\format{
A matrix containing three variables
\itemize{
\item x: x coordinate
\item y: y coordinate
\item t: t time
}
}
\usage{
camden_theft
}
\description{
A spatiotemporal point data
('theft-from-person' crime: Source: \code{police.uk})
of Camden Borough of London, UK
of the year 2021.
Note: Police.uk data is only accurate to monthly
scale (\code{yyyy-mm}). However, the data here is
provided at the daily scales
by adding random \code{days} to produce the dates in
the format \code{yyyy-mm-dd}. So, caution should be taken
when interpreting the results from this datasets.
}
\keyword{datasets}
