% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two spatiotemporal point patterns}
\usage{
compare(stp1, stp2, dim = "spatial")
}
\arguments{
\item{stp1}{(list or dataframe) First spatio-temporal
data containing three
basic columns: \code{x} - eastings,
\code{y} - northing, and \code{t} - time of occurrence
(in the format: `yyyy-mm-dd').}

\item{stp2}{(list or dataframe) Second spatio-temporal
data containing three
basic columns: \code{x} - eastings,
\code{y} - northing, and \code{t} - time of occurrence
(in the format: `yyyy-mm-dd').}

\item{dim}{(string) The dimension of the datasets
to compare. Default: \code{spatial}. Other available
options are: \code{temporal} and \code{spatiotemporal}.}
}
\value{
Returns series of graphs and statistics
describing the similarities between two spatiotemporal
datasets.
}
\description{
Generates series of visualizations
comparing two spatiotempora point patterns
}
\references{
https://www.google.co.uk/
}
