% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_areas.R
\name{compare_areas}
\alias{compare_areas}
\title{Compare two areas}
\usage{
compare_areas(area1, area2,
display_output = FALSE)
}
\arguments{
\item{area1}{(as \code{spatialPolygons},
\code{spatialPolygonDataFrames}, or
\verb{simple features}). the polygon object of the
first area.}

\item{area2}{(as \code{spatialPolygons},
\code{spatialPolygonDataFrames}, or
\verb{simple features}). the polygon object of the
second area.}

\item{display_output}{(logical) Whether to show output - a plot
comparing the two areas (plotted at the same scale), and a
statement stating the comparison.
Default: \code{FALSE}}
}
\value{
Returns a plot and a text (string) comparing
the sizes of two areas.
}
\description{
To compare the sizes of two
areas (boundary shapefiles).
}
\details{
Compares the sizes of two areas (polygon shapefiles).
The two shapefiles can be in any \code{crs},
and any spatial object formats. If enabled, the output
is shown comprising a plot and a string quoting the
factor by which
one area is larger than the other. This factor can
be used to scale specific spatial parameters, such as
\code{n_orign}, \code{s_threshold}, and \code{step_length}, during simulation
for a new study area.
}
\examples{
#load 'area1' object - boundary of Camden, UK
load(file = system.file("extdata", "camden.rda",
package="stppSim"))
camden_boundary = camden$boundary

#load 'area2' - boundary of Birmingham, UK
load(file = system.file("extdata", "birmingham_boundary.rda",
package="stppSim"))

#run
compare_areas(area1 = camden_boundary,
area2 = birmingham_boundary, display_output = FALSE)
}
