% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_spo.R
\name{constrained_spo}
\alias{constrained_spo}
\title{Simulate spatial point origins constrained
by the social configuration of the urban space.}
\usage{
constrained_spo(bpoly, p_ratio = 5,
show.plot=FALSE)
}
\arguments{
\item{bpoly}{(a spatialPolygonDataFrames) with binary attribute
field \code{class} with values \code{1} and \code{0}, representing
the enabled and disabled origins.}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio. For example, for a \code{20:80}
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Valid inputs are \code{10}, \code{20},
\code{30}, \code{40}, and \code{50}. A \code{30:70}, represents
30\% dominant and 70\% non-dominant origins.}

\item{show.plot}{(TRUE or FALSE) To display plot showing
points (origins).}
}
\value{
Returns the global temporal pattern
}
\description{
Simulate event origins (EOs) on a land use map
(contrained space) with binary classes \code{1} and \code{0}, representing
enabled and disabled origins. An enabled origin can
generate events while disabled origins can not generate
events. Each enabled origin is assigned
a probability value (representing the intensity) at which
the origin generates events in accordance with a specified
Pareto ratio.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
