% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_generator.R
\name{crp}
\alias{crp}
\title{Event (point) generators}
\usage{
crp(poly, npoints)
}
\arguments{
\item{poly}{(a list or dataframe) A list of boundary coordinates
(cartesian or geographic) of a study area}

\item{npoints}{(an integer) Number of points to generate}
}
\value{
Returns the global temporal pattern
}
\description{
Generates completely random points inside
a polygon boundary and assign probability values (signifying
their strength) in accordance with
the pareto principle (i.e. 80\% of the events
come from 20\% of the locations).
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
