% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_generator.R
\name{e_generator}
\alias{e_generator}
\title{Event (point) generators}
\usage{
e_generator(poly, npoints, p_ratio)
}
\arguments{
\item{poly}{(a list or dataframe) A list of boundary coordinates
(cartesian or geographic) of a study area}

\item{npoints}{(an integer) Number of points to generate}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio. For example, for a 20:80
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Input values must be \code{10}, \code{20},
\code{30}, \code{40}, or \code{50}. The 'p_ratio'
determines the proportion of points that are the most
dominant event generators.}
}
\value{
Returns the global temporal pattern
}
\description{
Generates completely random points inside
a polygon boundary and assign probability values (signifying
their strength) in accordance with
the pareto principle (i.e. 80\% of the events
come from 20\% of the locations).
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
