% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coords.R
\name{extract_coords}
\alias{extract_coords}
\title{Coordinates extraction}
\usage{
extract_coords(poly)
}
\arguments{
\item{poly}{(An sf or S4 object)
A polygon shapefile.}
}
\value{
Returns 2-column xy coordinates
}
\description{
Extracts the bounding (edges) coordinates
of a polygon object.
}
\details{
Given a spatial polygon object,
the function extracts the bounding coordinates
of the object.
}
\examples{
data(camden_boundary)
extract_coords(poly=camden_boundary)
}
