% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtp.R
\name{gtp}
\alias{gtp}
\title{Global temporal pattern (GTP)}
\usage{
gtp(start_date, trend = "stable",
slope = NULL, repeatType = "cyclic",
Tperiod = NULL, show.plot =FALSE)
}
\arguments{
\item{start_date}{the start date of the temporal pattern.
The date should be in the format \code{"yyyy-mm-dd"}.
The GTP will normally cover a 1-year period.}

\item{trend}{specifies the direction of the
long-term trend. Options are:
\code{"falling"}, \code{"stable"},
and \code{"rising"}. Default value is: \code{"stable"}.}

\item{slope}{slope of the long-term trend when
an \code{"rising"} or \code{"falling"} trend is specified.
Options: \code{"gentle"} or \code{"steep"}. The default value is
set as \code{NULL} for the \code{stable} trend.}

\item{repeatType}{type of short- to medium-term
fluctuations (patterns) associated with the
trend line. Options are: \code{`"cyclic"` and `"single"`}.
Default is: \code{`"cyclic"`}.}

\item{Tperiod}{time interval (in days) of repeat patterns
in the event occurrences. Default value is \code{90}
indicating the first seasonal
peak of cyclic repeat type. For \code{"single"} repeat type,
a single value, e.g. 90, or a sequence of values,
e.g. c(80:95), can be supplied.}

\item{show.plot}{(logical) Shows GTP.
Default is \code{FALSE}.}
}
\value{
Returns a time series (list) of 365
data points representing
1-year global temporal pattern.
}
\description{
Models the global temporal pattern,
as combining the long-term trend and seasonality.
}
\details{
Models the GTP for anchoring the temporal
trends and patterns of the point patterns to be simulated.
}
\examples{
gtp(start_date = "2020-01-01", trend = "stable",
slope = NULL, repeatType = "cyclic",
Tperiod = 60, show.plot = FALSE)
}
