% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landuse}
\alias{landuse}
\title{Landuse shapefile}
\format{
A boundary file (ESRI format)
\itemize{
\item type: Landuse type
\item rValues1: Numerical field setting a uniform
restriction value for each feature (Value \code{1} is set).
\item rValues2: Numerical field containing varying restriction
values for different feature classes.
}
}
\usage{
landuse
}
\description{
A land use shapefile of Camden
Borough of London, United Kingdom
}
\keyword{datasets}
