% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines}
\alias{lines}
\title{line or network data}
\format{
A simple feature with two fields
\itemize{
\item id: id of segment
\item geometry: geometry field
}
}
\usage{
lines
}
\description{
Two linestring representing
segments in a network
}
\keyword{datasets}
