% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grids.R
\name{make_grids}
\alias{make_grids}
\title{Make Square Grids System}
\usage{
make_grids(poly, size = 250,
show.output = FALSE, dir=NULL)
}
\arguments{
\item{poly}{(as \code{spatialPolygons}, \code{spatialPolygonDataFrames}, or
\verb{simple features}). A spatial polygon over
which the spatial grid is to be overlaid. Needs to be in a
cartesian CRS.}

\item{size}{Square grid size to be generated.
To be in the same unit associated with the \code{poly} (e.g.
metres, feets, etc.). Default: \code{200}.}

\item{show.output}{(logical) To show the output.
Default: \code{FALSE}}

\item{dir}{(character) Specifies the directory to
export the output. Default is \code{NULL}, indicating the
current working directory (cwd). A user can specify a different
directory in the format: "C:/.../folder".}
}
\value{
Returns a spatial square grid system
in a shapefile format
}
\description{
Generates a system of square grids over a specified
spatial boundary.
}
\references{
https://www.google.co.uk/
}
