% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poly}
\alias{poly}
\title{Boundary coordinates}
\format{
A dataframe containing one variable:
\itemize{
\item x: x coordinate
\item y: y coordinate
}
}
\usage{
poly
}
\description{
Boundary coordinates of Camden
Borough of London
}
\keyword{datasets}
