% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim.R
\name{psim}
\alias{psim}
\title{Modeling of the Global Temporal Pattern}
\usage{
psim(
  n_events,
  spo,
  s_threshold = 50,
  st_skewness = 0.5,
  ...,
  show.data = FALSE
)
}
\arguments{
\item{n_events}{(integer) Total Number of events to simulate.}

\item{spo}{(a list or dataframe) A list of spatial boundary
coordinates (or shapefile) within which the events are confined.
Should be generated using \code{random_spo} or \code{constrained_spo}
function.}

\item{s_threshold}{(numeric) Spatial threshold value. The
(assumed) spatial range within which events are
re-generated (or repeated) by or around the same origin.
Default: \code{250} (in the same linear unit as the \code{poly})
Default:\code{"daily"}. Other values are:
\code{"weekly", "monthly"}.}

\item{st_skewness}{(numeric) The tightness of events in space and time.
The value ranges from \code{0 - 1}, with event
volume being skewed towards the dominant origins, as the value tends
to \code{1}. Default: \code{0.5}. This index also controls the
total volume of events across space and time.}

\item{...}{additional arguments to pass from both the
\code{gtp} and \code{walker} functions. The latter is
utilized to define the properties of all the event
generators across the landscape.}

\item{show.data}{(TRUE or FALSE) To show the output data
Default is \code{FALSE}.}
}
\value{
Returns the global temporal pattern
}
\description{
Models the global temporal pattern (of
the point process) as consisting of the global linear
trend and the seasonality.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
