% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim_artif.R
\name{psim_artif}
\alias{psim_artif}
\title{Stpp from synthetic origins}
\usage{
psim_artif(n_events=1000, start_date = "yyyy-mm-dd",
poly, n_origin, resistance_feat, field,
n_foci, foci_separation, conc_type = "dispersed",
p_ratio, s_threshold = 50, step_length = 20,
trend = "stable", first_pDate=NULL,
slope = NULL, ..., show.plot=FALSE, show.data=FALSE)
}
\arguments{
\item{n_events}{(integer) Number of points
(events) to simulate. Default: \code{1000}.
A vector of integer values can be supplied, in the
format \code{c(a1, a2, ....)}, where a1, a2, ...
represent different values.}

\item{start_date}{The start date of temporal pattern.
The date should be in the format \code{"yyyy-mm-dd"}.
The GTP will usually covers a 1-year period.}

\item{poly}{(An sf or S4 object)
A polygon shapefile within which
event origins are to be situated.}

\item{n_origin}{(an integer) Number of locations from which
the walkers originate. Default:\code{50}.}

\item{resistance_feat}{(An S4 object) Optional
shapefile representing spaces across landscape
within which event}

\item{field}{A number in the range of \code{[0-1]}
(i.e. resistance values) to
assign to all features covered by \code{resistance_feat}; or
the name of a numeric field to extract such
resistance values for different feature classes.
The resistance value \code{0} and \code{1} indicate the
lowest and the highest restrictions, respectively,
to an event occuring within the space occupied
by a feature.
origins are not allowed. Default: \code{NULL}.}

\item{n_foci}{(an integer) Value indicating the number of
focal points amongst event origins. \code{n_foci} will usually be
smaller than \code{n_origin}.}

\item{foci_separation}{(an integer) A percentage value indicating
indicating the nearness of focal points from one another. A \code{0}
separation indicates that focal points are in close proximity
of one another, while a \code{100} indicates focal points being
evenly distributed across space.}

\item{conc_type}{(string) Specifies the spatial pattern
of non-focal origin (strengths) in relation to
to their nearest focal origins. Value is either
\code{"nucleated"} or \code{"dispersed"}.}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio.
For example, a value of \code{20}
implies a \code{20:80} Pareto ratio.}

\item{s_threshold}{(numeric) Spatial threshold
value. This is the spatial range within which
a walker perceives it's environment at any
instant. Default: \code{250} (in the same linear unit
as the \code{poly} - polygon shapefile).}

\item{step_length}{(numeric) A maximum step taken at a time
by a walker from one point to the next.}

\item{trend}{(string) Specify the trend direction of
the GTP. Values are: \code{"decreasing"}, \code{"stable"},
and \code{"increasing"}. Default is: \code{"stable"}.}

\item{first_pDate}{(in \code{"yyyy-mm-dd"} format).
Date of the
first seasonal peak of the time series.
Default value is \code{NULL}, in which a
seasonal cycle of 180 days is utilized. That is,
a first seasonal peak of 90 days.}

\item{slope}{(string) Slope GTP trend if
"increasing" or "decreasing" trend is specified.
Values: \code{"gentle"} or \code{"steep"}.
Default value is \code{NULL} (i.e., for \code{stable} trend).}

\item{...}{additional arguments to pass from
\code{gtp}, \code{walker} and \code{artif_spo}
functions.}

\item{show.plot}{(logical) Shows GTP.
Default is \code{FALSE}.}

\item{show.data}{(TRUE or FALSE) To show the output
data. Default is \code{FALSE}.}
}
\value{
Returns a list of artificial spatiotemporal
point patterns.
}
\description{
Generate spatiotemporal
point pattern from synthesized origins.
}
\details{
Generate spatiotemporal point pattern
based on the actions of specified 'walkers' moving
across a landscape. Both the walkers and the
landscape are configured arbitrarily (in accordance
with the users (expert) knowledge of the domain
in question.
}
\examples{
\dontrun{
data(camden_boundary)
data(landuse)
artif_stpp <- psim_artif(n_events=200, start_date = "2021-01-01",
poly=camden_boundary, n_origin=50, resistance_feat = landuse,
field = "rValue1",
n_foci=5, foci_separation = 10, conc_type = "dispersed",
p_ratio = 20, s_threshold = 50, step_length = 20,
trend = "stable", first_pDate=NULL,
slope = NULL,show.plot=FALSE, show.data=FALSE)
}
}
