% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim_real.R
\name{psim_real}
\alias{psim_real}
\title{Point Pattern Simulation from real origins}
\usage{
psim_real(n_events, ppt, start_date = NULL, poly = NULL,
n_origin=50, p_ratio=20, crsys = "CRS_string")
}
\arguments{
\item{n_events}{(integer) Value of the total
number of points (events) to simulate. Default: \code{2000}.
A vector of integer values can also be inputted, such as
\code{c(a1, a2, ....)}, where a1, a2, ... represent different values.}

\item{ppt}{(matrix) coordinates (and time) vectors of points.
A 3-column matrix or list: \code{x} - eastings,
\code{y} - northing, and \code{t} - time of occurrence
(in the format: `yyyy-mm-dd'). The sample needs to
cover the one year period being investigated.}

\item{start_date}{(string) Specify the start date of
the sample point supplied (i.e. \code{ppt}),
in the format \code{yyyy-mm-dd}. If \code{NULL}, the earliest date
from the \code{ppt} is utilized. The end date is the 365th day
from the specified start date (by default).}

\item{poly}{(as \code{spatialPolygons},
\code{spatialPolygonDataFrames}, or
\verb{simple features}). The boundary (spatial polygon) surrounding
the sample points. The default is \code{NULL} - meaning that an
arbitrary boundary is drawn to cover the spatial
point distribution. The 'poly' object must have a projection
system (crs) when not NULL.}

\item{n_origin}{(an integer) Number of origins to simulate.
Default:\code{50}. This is the parameter that has the greatest
influence on the computational time.}

\item{p_ratio}{(an integer) The smaller of the
two terms of the Pareto ratio. For example, for a \code{20:80}
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Valid inputs are \code{10}, \code{20},
\code{30}, \code{40}, and \code{50}. A \code{30:70}, represents
30\% dominant and 70\% non-dominant origins.}

\item{crsys}{(string) The projection ('crs') system to utilize
when 'poly' argument is not NULL. You can obtain CRS string
from "http://spatialreference.org/". The \code{crs} can be set using
\verb{proj4string(poly) <- "CRS string", where }CRS string\verb{defines the projection of}ppt\verb{. When both }poly\code{and}crsys`
are not NULL, the function utilizes the crs of the former}
}
\description{
Generate point pattern in space and time
from real sample dataset, based on the
temporal pattern and
spatial properties learnt from the sample dataset.
}
\details{
Returns an object of the class \code{real_spo},
detailing the spatiotemporal properties of a real
sample dataset
}
\examples{
\dontrun{
data(SanF_fulldata)
data(SanF_CRS_string)
#get a sample data
set.seed(1000)
sample_size <- 1000
dat_sample <- SanF_fulldata[sample(1:nrow(SanF_fulldata),
sample_size, replace=FALSE),]
result <- psim_real(n_events=2000, ppt=dat_sample,
start_date = NULL, poly = NULL,
n_origin=50, p_ratio=20,crsys = SanF_CRS_string)
}
}
\references{
https://www.google.co.uk/
}
