% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pts}
\alias{pts}
\title{point data}
\format{
A simple feature with two fields
\itemize{
\item id: point id
\item geometry: geometry field
}
}
\usage{
pts
}
\description{
Point patterns
}
\keyword{datasets}
