% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_EO.R
\name{random_EO}
\alias{random_EO}
\title{Simulate random event origins}
\usage{
random_EO(poly, npoints, p_ratio, show.plot=FALSE)
}
\arguments{
\item{poly}{(a list or dataframe) A list of boundary coordinates
(cartesian or geographic).}

\item{npoints}{(an integer) Number of points to simulate}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio. For example, for a \code{20:80}
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Valid inputs are \code{10}, \code{20},
\code{30}, \code{40}, and \code{50}. A \code{30:70}, represents
30\% dominant and 70\% non-dominant origins.}

\item{show.plot}{(TRUE or FALSE) To display plot showing
points (origins).}
}
\value{
Returns the global temporal pattern
}
\description{
Simulate completely random points inside
a polygon boundary and assign probability values (signifying
their strength) in accordance with a specified
Pareto ratio. The points represent the origin
from which events originate.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
