% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_spo.R
\name{random_spp}
\alias{random_spp}
\title{Simulate random spatial point patterns}
\usage{
random_spp(poly, npoints, p_ratio, show.plot=FALSE)
}
\arguments{
\item{poly}{(a list or dataframe) A list of spatial boundary
coordinates within which the events are confined.}

\item{npoints}{(an integer) Number of origins (points) to simulate}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio. For example, for a \code{20:80}
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Valid inputs are \code{10}, \code{20},
\code{30}, \code{40}, and \code{50}. A \code{30:70}, represents
30\% dominant and 70\% non-dominant origins.}

\item{show.plot}{(TRUE or FALSE) To display plot showing
points (origins).}
}
\value{
Returns the global temporal pattern
}
\description{
Simulate spatial point pattern from
randomly created event origin. Each origin is assigned
a probability value (representing the intensity) at which
the origin generates events in accordance with a specified
Pareto ratio.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
