% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regular_poly}
\alias{regular_poly}
\title{A rectangular boundary coordinates}
\format{
A dataframe containing one variable:
\itemize{
\item x: x coordinate
\item y: y coordinate
}
}
\usage{
regular_poly
}
\description{
A rectangular boundary coordinates
}
\keyword{datasets}
