% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{San_Francisco}
\alias{San_Francisco}
\title{A land use shapefile of a portion of
San Francisco City, United States}
\format{
A boundary file (ESRI format)
\itemize{
\item landuse_1: land use categories denoting
the social configuration of the urban space
\item class: a binary field indicating origins that
have the ability to generate events ('1') and origins
that lack the ability to generate points ('0').
}
}
\usage{
San_Francisco
}
\description{
A land use shapefile of a portion of
San Francisco City, United States
}
\keyword{datasets}
