% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SanF_CRS_string}
\alias{SanF_CRS_string}
\title{CRS string of San Francisco Area
Source: http://spatialreference.org/}
\format{
A list containing projected CRS
string for San Francisco area
\itemize{
\item text: CRS string
}
}
\usage{
SanF_CRS_string
}
\description{
CRS string of San Francisco Area
Source: http://spatialreference.org/
}
\keyword{datasets}
