% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SanF_fulldata}
\alias{SanF_fulldata}
\title{A complete point data of a portion of
San Francisco city.}
\format{
A matrix containing two variables
\itemize{
\item x: x coordinate
\item y: y coordinate
\item t: t time
}
}
\usage{
SanF_fulldata
}
\description{
A complete point data of a portion of
San Francisco city.
}
\keyword{datasets}
