% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SanF_sample}
\alias{SanF_sample}
\title{A sample point data of a part of San Francisco city}
\format{
A matrix containing two variables
\itemize{
\item x: x coordinate
\item y: y coordinate
}
}
\usage{
SanF_sample
}
\description{
A sample point data of a part of San Francisco city
}
\keyword{datasets}
