% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_resistance.R
\name{space_resistance}
\alias{space_resistance}
\title{Space (landscape) resistance raster}
\usage{
space_resistance(shp, baseRaster, res, binary = is.na(field),
field = NA, mapvalues = NA, extend = TRUE)
}
\arguments{
\item{shp}{either a character string specifying
the shapefile filename or a shapefile object itself.}

\item{baseRaster}{if provided, a raster onto which
to stack the given rasterized shapefile.}

\item{res}{the desired pixel resolution of the
raster to be created, when baseRaster is not provided.}

\item{binary}{if TRUE, the shapefile will be rasterized
so that any feature is assigned a value of 0,
and the background 1.}

\item{field}{either a number in the range 0 to 1, in
which case it will be assigned to all pixels covered
by features of the shapefile; or the name of the numeric
shapefile field from which to extract such number; or
the name of the factor shapefile field containing classes
to map to resistance values (see mapvalues).}

\item{mapvalues}{a named vector specifying the resistance
value mapping from the classes of field.}

\item{extend}{set to TRUE to extend baseRaster if the shapefile
has a larger extent. If FALSE, the shapefile will be clipped
to baseRaster's extent.}
}
\value{
Returns a trace of walker's path, and the
corresponding events.
}
\description{
Build a space resistance map
from one or more shapefiles. Function builds
on functions from package raster- and
SimRIv-packages.
}
\references{
\enumerate{
\item Paul Murrell (2019). rasterize: Rasterize Graphical
Output. R package version 0.1.
https://CRAN.R-project.org/package=rasterize
\item Quaglietta L, Porto M (2019). \emph{SiMRiv: Individual-Based,
Spatially-Explicit Simulation and Analysis of Multi-State
Movements in River Networks and Heterogeneous Landscapes}.
R package version 1.0.4, <URL:
https://CRAN.R-project.org/package=SiMRiv>.
}
}
