% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_pattern.R
\name{t_pattern}
\alias{t_pattern}
\title{Modelling of the global temporal pattern}
\usage{
t_pattern(start_date = "01-01", trend = "stable",
slope = "gentle", first_s_peak=90, scale = 1, show.plot =FALSE)
}
\arguments{
\item{start_date}{The start date of the study period.
Default value is \code{"01-01"} (i.e. January 1st). By default
the end date of the study period will be \code{"12-31"} (i.e.
31st December). A user can specify any start date in the
format \code{"mm/dd"}. The end date is the following 365th day
from the specified start date.}

\item{trend}{(a character) specifying the direction of
the global linear trend of the point process. Options
are \code{"decreasing"}, \code{"stable"},
and \code{"increasing"} linear trends, respectively. Default
value is \code{"stable"}.}

\item{slope}{(a character) specifying the type of the slope
if the trend is "increasing" or "decreasing". Options
are: \code{"gentle"} and \code{"steep"}. Default value is \code{"gentle"}.}

\item{first_s_peak}{Number of days before the first seasonal
peak. Default: \code{90}. This implies a seasonal cycle
of 180 days.}

\item{scale}{(an integer) For scaling point count}

\item{show.plot}{(TRUE or False) To show the time series
plot. Default is \code{FALSE}.}
}
\value{
Returns the global temporal pattern
}
\description{
Models the global temporal pattern (of
the point process) as consisting of the global linear
trend and the seasonality.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
