% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xyt_data}
\alias{xyt_data}
\title{Spatiotemporal point data}
\format{
A matrix containing three variables
\itemize{
\item x: x coordinate
\item y: y coordinate
\item t: t time
}
}
\usage{
xyt_data
}
\description{
Example spatiotemporal point data
of a part of San Francisco City,
California, US
}
\keyword{datasets}
