% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xyz}
\alias{xyz}
\title{xyz data}
\format{
A matrix containing three variables
\itemize{
\item x: x coordinate
\item y: y coordinate
\item z: z height/probability/etc
}
}
\usage{
xyz
}
\description{
Example data with 'x', 'y',
and a 'z' information
}
\keyword{datasets}
