% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artif_spo.R
\name{artif_spo}
\alias{artif_spo}
\title{Generate artificial (random) spatial point origins}
\usage{
artif_spo(poly, n_origin, p_ratio, show.plot=FALSE)
}
\arguments{
\item{poly}{(A dataframe or S4 object) A dataframe of X, Y
coordinates or a spatial boundary (as "SpatialPolygonsDataFrame",
"SpatialPolygons", or "sf") representing the boundary within which
events are to be generated.}

\item{n_origin}{(an integer) Number of origins to simulate.
Default:\code{50}. This is the parameter that has the greatest
influence on the computational time.}

\item{p_ratio}{(an integer) The smaller of the
two terms of the Pareto ratio. For example, for a \code{20:80}
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Valid inputs are \code{10}, \code{20},
\code{30}, \code{40}, and \code{50}. A \code{30:70}, represents
30\% dominant and 70\% non-dominant origins.}

\item{show.plot}{(TRUE or FALSE) To display plot showing
origins.}
}
\value{
Returns random event origins
}
\description{
Simulate artificial point origin across a
defined spatial boundary.Each origin is assigned
a probability value (representing the relative intensity) at which
the origin generates events with respect to a specified
Pareto ratio.
}
\details{
Returns an object of the class \code{artif_spo},
detailing the properties of artificial (simulated) spatial point
origins.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
