% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camden_crimes}
\alias{camden_crimes}
\title{Records of crimes of Camden Borough of London,
UK, 2021 (Source: https://data.police.uk/data/)}
\format{
A matrix containing four variables
\itemize{
\item x: x coordinate
\item y: y coordinate
\item date: date of occurence
\item type: types of crime
}
}
\usage{
camden_crimes
}
\description{
Data comprising 'Theft' and 'Criminal Damage'
records of Camden Borough of London, UK
for the year 2021 (Source: \verb{https://data.police.uk/}).
Note: Police.uk data is aggregated at monthly
scale (\code{yyyy-mm}). But, the data provided here has been
disaggregated to daily scale by adding fake
'daily' stamps (to give \code{yyyy-mm-dd}). So, caution should
be taken when interpreting the results based on
full date.
}
\keyword{datasets}
