% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camden_theft}
\alias{camden_theft}
\title{A spatiotemporal point data
('theft-from-person' crime)
of Camden Borough of London, UK
for the year 2021}
\format{
A matrix containing three variables
\itemize{
\item x: x coordinate
\item y: y coordinate
\item t: t time
}
}
\usage{
camden_theft
}
\description{
A spatiotemporal point data
('theft-from-person' crime)
of Camden Borough of London, UK
for the year 2021
}
\keyword{datasets}
