% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chull_poly.R
\name{chull_poly}
\alias{chull_poly}
\title{Generate boundary around a set of Points}
\usage{
chull_poly (xycoords,
crsys = NULL)
}
\arguments{
\item{xycoords}{(matrix) coordinate vectors of points.
A 2-column matrix or list: \code{x} - the eastings,
and \code{y} - the northing.}

\item{crsys}{(string) coordinate reference system of
the resulting polygon. e.g.,
"+proj=longlat +datum=WGS84".
Default: \code{NULL}.}
}
\description{
Generate a boundary (polygon) around
a set of sample points, using Convex Hull technique.
}
\details{
The algorithm is that given by Eddy (1977).
}
\examples{
data(SanF_fulldata)
SanF_fulldata_xy <-
matrix(as.numeric(SanF_fulldata[,1:2]),,2)
bry <- chull_poly(SanF_fulldata_xy, crsys = NULL)
}
\references{
Eddy, W. F. (1977).
A new convex hull algorithm for planar sets.
ACM Transactions on Mathematical Software,
3, 398--403.10.1145/355759.355766.
}
