% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_spo.R
\name{constrained_spo}
\alias{constrained_spo}
\title{Simulate spatial point origins constrained
by the landscape (use) configuration of the urban space.}
\usage{
constrained_spo(bpoly, p_ratio = 5,
show.plot=FALSE)
}
\arguments{
\item{bpoly}{(A dataframe or S4 object) A dataframe of X, Y
coordinates or a spatial boundary (as "SpatialPolygonsDataFrame",
"SpatialPolygons", or "sf") representing the boundary within which
events are to be generated. Must include a field named \code{class}
with entries \code{1's} and \code{0's} (i.e. binary) representing
the active and non-active origins, respectively.}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio. For example, for a \code{20:80}
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Valid inputs are \code{10}, \code{20},
\code{30}, \code{40}, and \code{50}. A \code{30:70}, represents
30\% dominant and 70\% non-dominant origins.}

\item{show.plot}{(TRUE or FALSE) To display plot showing
base map (i.e. social configuration of the landscape, in
terms of active and non-active spaces), and the event origins.}
}
\value{
Returns the event origins constraint by the
social configuration of the space
}
\description{
Simulate event origins (EOs) on a land use map
(contrained space) with binary classes \code{1} and \code{0}, representing
active and non-active origins. An \code{active} origin can
generate events while a \code{non-active} origin can not generate
events. Each active origin is assigned
a probability value (representing the intensity) at which
the origin generates events in accordance with a specified
Pareto ratio.
}
\details{
Note: The \code{class} field of 'bpoly'
will be utilized for mapping the basemap.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
