% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_checker.R
\name{date_checker}
\alias{date_checker}
\title{Date Format Checker}
\usage{
is_date(x, format="%Y-%m-%d")
}
\arguments{
\item{x}{A date input}
}
\value{
Returns TRUE or FALSE
}
\description{
Checks if date is in the correct format, i.e.
\code{yyyy-mm-dd}
}
\details{
Returns TRUE or FALSE, to indicate
that the date input is in the correct or incorrect
format, respectively
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
