% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_checker.R
\name{date_checker}
\alias{date_checker}
\title{Date Format Checker}
\usage{
date_checker(x, format=NULL)
}
\arguments{
\item{x}{A date input}

\item{format}{The date format required.
Default: NULL.}
}
\value{
Returns TRUE or FALSE
}
\description{
Checks if date is in the correct format, i.e.
\code{yyyy-mm-dd}
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
