% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coords.R
\name{extract_coords}
\alias{extract_coords}
\title{Extracting coordinates of a polygon boundary}
\usage{
extract_coords(poly)
}
\arguments{
\item{poly}{(a spatialPolygons, spatialPolygonDataFrames, or
an "sf").
The polygon object must be in a
Cartesian coordinate reference system.}
}
\value{
Returns the global temporal pattern
}
\description{
Given a polygon object, the goal is to extract
the coordinates of the edges of the boundary.
}
\references{
https://www.google.co.uk/
}
