% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtp.R
\name{gtp}
\alias{gtp}
\title{Global temporal pattern (GTP)}
\usage{
gtp(start_date, trend = "stable",
slope = NULL, first_pDate = NULL, show.plot =FALSE)
}
\arguments{
\item{start_date}{The start date of temporal pattern.
The date should be in the format \code{"yyyy-mm-dd"}.
The GTP will usually covers a 1-year period.}

\item{trend}{(string) Specify the trend direction of
the GTP. Values are: \code{"decreasing"}, \code{"stable"},
and \code{"increasing"}. Default is: \code{"stable"}.}

\item{slope}{(string) Slope GTP trend if
"increasing" or "decreasing" trend is specified.
Values: \code{"gentle"} or \code{"steep"}.
Default value is \code{NULL} (i.e., for \code{stable} trend).}

\item{first_pDate}{(in \code{"yyyy-mm-dd"} format).
Date of the
first seasonal peak of the time series.
Default value is \code{NULL}, in which a
seasonal cycle of 180 days is utilized. That is,
a first seasonal peak of 90 days.}

\item{show.plot}{(logical) Shows GTP.
Default is \code{FALSE}.}
}
\value{
Returns a vector of 365 data points representing
the global temporal pattern
}
\description{
Models the global temporal pattern,
as combining long-term trend and seasonality.
}
\details{
Models the GTP which could be utilize for simulating
artifical point pattern across space.
}
\examples{
gtp(start_date = "2020-01-01", trend = "stable",
slope = NULL, first_pDate = "2020-02-28", show.plot = FALSE)
}
