% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtp.R
\name{gtp}
\alias{gtp}
\title{Modeling of the Global Temporal Pattern}
\usage{
gtp(start_date = "2000-01-01", trend = "stable",
slope = "NULL", first_s_peak=as.Date(start_date)+90, show.plot =FALSE)
}
\arguments{
\item{start_date}{(Date) The start date of simulation. The date should
be in the format \code{"yyyy-mm-dd"}. Default value is
\code{"2000-01-01"}. A specified date can be earlier or later
than this stated default value. By default, a 1-year worth of
date is simulated. In other words, the end date of
simulation is the next 365th day
from the specified start date.}

\item{trend}{(a character) Specifying the direction of
the global (linear) trend of the simulated time series.
Three options
available are \code{"decreasing"}, \code{"stable"},
and \code{"increasing"} trends. Default: \code{"stable"}.}

\item{slope}{(a character) Slope angle for an
"increasing" or "decreasing" trend. Two options
are available: \code{"gentle"} and \code{"steep"}.
Default value is \code{"NULL"} for the default trend
(i.e. \code{stable}).}

\item{first_s_peak}{(Date) The date that marks the
first seasonal peak of the time series.
Default value is \code{as.Date(start_date)+90}, i.e. 90 days after the
specified \code{start_date} (implying a seasonal cycle of
180 days. The date should
be in the format: \code{"yyyy-mm-dd"}.}

\item{show.plot}{(TRUE or False) To show the time series
plot. Default is \code{FALSE}.}
}
\value{
Returns the global temporal pattern
}
\description{
Models the global temporal pattern (of
the point process) as consisting of the global linear
trend and the seasonality.
}
\details{
Returns an object of the class \code{artif_gtp},
describing an artificial global temporal patterns.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
