% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtp.R
\name{gtp}
\alias{gtp}
\title{Modeling of the Global Temporal Pattern}
\usage{
gtp(start_date = "01-01", t_resolution = 1, trend = "stable",
slope = "NULL", first_s_peak=90, show.plot =FALSE)
}
\arguments{
\item{start_date}{The start date of the study period.
Default value is \code{"01-01"} (i.e. January 1st). By default
the end date of the study period is set as \code{"12-31"} (i.e.
31st December). A user can specify any start date in the
format \code{"mm/dd"}. The end date is the next 365th day
from the specified start date.}

\item{t_resolution}{(character) The temporal resolution
at which events are re-generated (or repeated). Specified in
number of days. Default:\code{1} (currently the only option
available).}

\item{trend}{(a character) Specifying the direction of
the global (linear) trend of the point process. Three options
available are \code{"decreasing"}, \code{"stable"},
and \code{"increasing"} trends. Default: \code{"stable"}.}

\item{slope}{(a character) Slope angle for an
"increasing" or "decreasing" trend. Two options
are available: \code{"gentle"} and \code{"steep"}.
Default value is \code{"NULL"} for the default trend
(i.e. \code{stable}).}

\item{first_s_peak}{Number of days before the first seasonal
peak. Default: \code{90}. This implies a seasonal cycle}

\item{show.plot}{(TRUE or False) To show the time series
plot. Default is \code{FALSE}.}
}
\value{
Returns the global temporal pattern
}
\description{
Models the global temporal pattern (of
the point process) as consisting of the global linear
trend and the seasonality.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
