% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landuse}
\alias{landuse}
\title{Landuse shapefile}
\format{
A boundary file (ESRI format)
\itemize{
\item type: Landuse type
\item rValues1: Field specifying a uniform resistance
for all features (Value = 1)
for all features
\item rValues2: Field specifying varying resistance
values for different feature classes.
}
}
\usage{
landuse
}
\description{
A land use shapefile of Camden
Borough of London, United Kingdom
}
\keyword{datasets}
