% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grids.R
\name{make_grids}
\alias{make_grids}
\title{Make square grids}
\usage{
make_grids(poly, size = 250,
show_output = FALSE)
}
\arguments{
\item{poly}{(as \code{spatialPolygons},
\code{spatialPolygonDataFrames}, or
\verb{simple features}). A polygon object over
which square grids are to be overlaid.}

\item{size}{Size of square grids to be
created. For example, the input \code{size}
for a 250 by 250 square grids is \code{250}.}

\item{show_output}{(logical) Display the output.
Default: \code{FALSE}}
}
\value{
Returns a spatial square grid system
in a shapefile format
}
\description{
Generates a system of square grids
across a specified spatial polygon (boundary).
}
\details{
Exports a grid system in a shapefile
format (in the same crs as the input \code{poly})
}
\examples{
data(camden_boundary)
make_grids(poly=camden_boundary, size = 250,
show_output = FALSE)
}
