% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_prob.R
\name{p_prob}
\alias{p_prob}
\title{Pareto Probability distribution}
\usage{
p_prob(npoints,  p_ratio = 30)
}
\arguments{
\item{npoints}{(an integer) Number of points. Default is
\code{50}.}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio. For example, for a 20:80
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Input values must be \code{10}, \code{20},
\code{30}, \code{40}, or \code{50}. The 'p_ratio'
determines the proportion of points that are the most
dominant event generators.}
}
\value{
Returns the global temporal pattern
}
\description{
Given a specified number of points \code{n},
this function generates an \code{n} probability values
in accordance with the Pareto principle.
}
\references{

}
