% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_prob.R
\name{p_prob}
\alias{p_prob}
\title{Pareto Probability distribution}
\usage{
p_prob(n_origin,  p_ratio = 30)
}
\arguments{
\item{n_origin}{(an integer) Number of origins. Default is
\code{50}.}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio. For instance, for a
\code{20:80} ratio, \code{p_ratio} will be \code{20}. Default value is
\code{20}. Input values must be \code{5}, \code{10}, \code{20},
\code{30}, or \code{40}. The 'p_ratio'
determines the proportion of points that are the most
dominant event generators.}
}
\value{
Returns the global temporal pattern
}
\description{
Given a specified number of points \code{n},
this function generates an \code{n} probability values
in accordance with a specified Pareto ratio.
}
\references{
https://www.google.co.uk/
}
