% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_prob.R
\name{p_prob}
\alias{p_prob}
\title{Pareto Probability distribution}
\usage{
p_prob(npoints)
}
\arguments{
\item{npoints}{(an integer) Number of points}
}
\value{
Returns the global temporal pattern
}
\description{
Given a number of points, this function
assign probability values (signifying
their strength) to the points in accordance with
the pareto principle (i.e. 80\% of the events
come from 20\% of the locations).
}
\references{

}
