% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_tester.R
\name{poly_tester}
\alias{poly_tester}
\title{Test the geometry and CRS of a polygon}
\usage{
poly_tester(poly)
}
\arguments{
\item{poly}{(as \code{spatialPolygons}, \code{spatialPolygonDataFrames}, or
\verb{simple features}). A spatial polygon representing a
landscape coverage.}
}
\value{
Returns error messages if
the polygon is not in correct geometry
or CRS.
}
\description{
To test that a polygon has the correct
geometry as well as a linear or Cartesian CRS.
}
\references{
#https://google.co.uk
}
