% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim.R
\name{psim}
\alias{psim}
\title{Modeling of the Global Temporal Pattern}
\usage{
gtp(start_date = "01-01", trend = "stable",
slope = "NULL", first_s_peak=90, scale = 1, show.plot =FALSE)
}
\arguments{
\item{start_date}{The start date of the study period.
Default value is \code{"01-01"} (i.e. January 1st). By default
the end date of the study period is set as \code{"12-31"} (i.e.
31st December). A user can specify any start date in the
format \code{"mm/dd"}. The end date is the next 365th day
from the specified start date.}

\item{trend}{(a character) Specifying the direction of
the global (linear) trend of the point process. Three options
available are \code{"decreasing"}, \code{"stable"},
and \code{"increasing"} trends. Default: \code{"stable"}.}

\item{slope}{(a character) Slope angle for an
"increasing" or "decreasing" trend. Two options
are available: \code{"gentle"} and \code{"steep"}.
Default value is \code{"NULL"} for the default trend
(i.e. \code{stable}).}

\item{first_s_peak}{Number of days before the first seasonal
peak. Default: \code{90}. This implies a seasonal cycle
of 180 days.}

\item{scale}{(an integer) For scaling point counts. Default: \code{1}}

\item{poly}{(a list or dataframe) A list of spatial boundary
coordinates within which the events are confined.}

\item{npoints}{(an integer) Number of origins (points) to simulate}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio. For example, for a \code{20:80}
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Valid inputs are \code{10}, \code{20},
\code{30}, \code{40}, and \code{50}. A \code{30:70}, represents
30\% dominant and 70\% non-dominant origins.}

\item{show.plot}{(TRUE or False) To show the time series
plot. Default is \code{FALSE}.}
}
\value{
Returns the global temporal pattern
}
\description{
Models the global temporal pattern (of
the point process) as consisting of the global linear
trend and the seasonality.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
