% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim.R
\name{psim}
\alias{psim}
\title{Point Pattern Simulation}
\usage{
psim(
  n_events = 2000,
  start_date = "2000-01-01",
  spo,
  s_threshold = 50,
  st_skewness = 0.5,
  trend = "stable",
  first_s_peak,
  slope = NULL,
  ...,
  show.data = FALSE
)
}
\arguments{
\item{n_events}{(integer) Value of the total
number of points (events) to simulate. Default: \code{2000}.
A vector of integer values can also be inputted, such as
\code{c(a1, a2, ....)}, where a1, a2, ... represent different values.}

\item{start_date}{(Date) The start date of simulation. The date should
be in the format \code{"yyyy-mm-dd"}. Default value is
\code{"2000-01-01"}. A specified date can be earlier or later
than this stated default value. By default, a 1-year worth of
date is simulated. In other words, the end date of
simulation is the next 365th day
from the specified start date.}

\item{spo}{(a list or dataframe) A list of spatial boundary
coordinates (or shapefile) within which the events are confined.
Should be generated using \code{random_spo} or \code{constrained_spo}
function. The \code{spo$poly} output is set as \code{poly} argument
in this function.}

\item{s_threshold}{(numeric) Spatial threshold value. The
(assumed) spatial range within which events are
re-generated (or repeated) by or around the same origin.
Default: \code{250} (in the same linear unit as the \code{poly})}

\item{st_skewness}{(numeric) The tightness of events in space and time.
The value ranges from \code{0 - 1}, with event
volume being skewed towards the dominant origins, as the value tends
to \code{1}. Default: \code{0.5}. This index also controls the
total volume of events across space and time.}

\item{trend}{(a character) Specifying the direction of
the global (linear) trend of the simulated time series.
Three options
available are \code{"decreasing"}, \code{"stable"},
and \code{"increasing"} trends. Default: \code{"stable"}.}

\item{first_s_peak}{(Date) The date that marks the
first seasonal peak of the time series.
Default value is \code{as.Date("2000-01-01")+90},
i.e. 90 days after the
specified \code{start_date} (implying a seasonal cycle of
180 days. The date should
be in the format: \code{"yyyy-mm-dd"}.}

\item{slope}{(a character) Slope angle for an
"increasing" or "decreasing" trend. Two options
are available: \code{"gentle"} and \code{"steep"}.
Default value is \code{"NULL"} for the default trend
(i.e. \code{stable}).}

\item{...}{additional arguments to pass from
\code{gtp} and \code{walker} functions. Arguments from
\code{gtp} can be used to define the nature of the
temporal trend and pattern over time, while arguments
from \code{walker} can be utilized to define the properties
of event generators (walkers) across the landscape.}

\item{show.data}{(TRUE or FALSE) To show the output data
Default is \code{FALSE}.}
}
\value{
Returns the global temporal pattern
}
\description{
Generate point pattern in space and time
based on a specified global temporal pattern and
spatial properties.
}
\details{
A list containing \code{k} entries,
where \code{k} is the length of \code{n_events} (see above).
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
