% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim.R
\name{psim}
\alias{psim}
\title{Modeling of the Global Temporal Pattern}
\usage{
psim(
  spo,
  start_date,
  s_threshold = 50,
  step_length = 20,
  show.data = TRUE,
  trend,
  slope,
  first_s_peak,
  npoints,
  p_ratio
)
}
\arguments{
\item{spo}{(a list or dataframe) A list of spatial boundary
coordinates (or shapefile) within which the events are confined.
Should be generated using \code{random_spo} or \code{constrained_spo}
function.}

\item{start_date}{The start date of the study period.
Default value is \code{"01-01"} (i.e. January 1st). By default
the end date of the study period is set as \code{"12-31"} (i.e.
31st December). A user can specify any start date in the
format \code{"mm/dd"}. The end date is the next 365th day
from the specified start date.}

\item{s_threshold}{(numeric) Spatial threshold value. The
(assumed) spatial range within which events are
re-generated (or repeated) by or around the same origin.
Default: \code{250} (in the same linear unit as the \code{poly})
Default:\code{"daily"}. Other values are:
\code{"weekly", "monthly"}.}

\item{step_length}{(numeric) A maximum step taken at a time
by a walker from one state to the next. Should be a fraction
of the spatial units of the landscape. Default: half the size
of the minimum spatial unit in a landscape
(for a constraint landscape) or}

\item{show.data}{(TRUE or FALSE) To show the output data
Default is \code{FALSE}.}

\item{trend}{(a character) Specifying the direction of
the global (linear) trend of the point process. Three options
available are \code{"decreasing"}, \code{"stable"},
and \code{"increasing"} trends. Default: \code{"stable"}.}

\item{slope}{(a character) Slope angle for an
"increasing" or "decreasing" trend. Two options
are available: \code{"gentle"} and \code{"steep"}.
Default value is \code{"NULL"} for the default trend
(i.e. \code{stable}).}

\item{first_s_peak}{Number of days before the first seasonal
peak. Default: \code{90}. This implies a seasonal cycle
of 180 days.}

\item{npoints}{(an integer) Number of origins (points) to simulate}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio. For example, for a \code{20:80}
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Valid inputs are \code{10}, \code{20},
\code{30}, \code{40}, and \code{50}. A \code{30:70}, represents
30\% dominant and 70\% non-dominant origins.}
}
\value{
Returns the global temporal pattern
}
\description{
Models the global temporal pattern (of
the point process) as consisting of the global linear
trend and the seasonality.
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
