% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim_artif.R
\name{psim_artif}
\alias{psim_artif}
\title{Point Pattern Simulation from artificial origins}
\usage{
psim_artif(
  n_events = 2000,
  start_date = "yyyy-mm-dd",
  poly,
  n_origin,
  space_resist = space_resist,
  n_foci,
  foci_separation,
  p_ratio,
  s_threshold = 50,
  trend = "stable",
  first_pDate = NULL,
  slope = NULL,
  ...,
  show.plot = FALSE,
  show.data = FALSE
)
}
\arguments{
\item{n_events}{(integer) Value of the total
number of points (events) to simulate. Default: \code{2000}.
A vector of integer values can also be inputted, such as
\code{c(a1, a2, ....)}, where a1, a2, ... represent different values.}

\item{start_date}{(Date) The start date of simulation. The date should
be in the format \code{"yyyy-mm-dd"}. A specified date can be earlier or later
than this stated default value. By default, a 1-year worth of
date is simulated. In other words, the end date of
simulation is the next 365th day
from the specified start date.}

\item{poly}{(A dataframe or S4 object) A dataframe of X, Y
coordinates or a spatial boundary (as "SpatialPolygonsDataFrame",
"SpatialPolygons", or "sf") representing the boundary within which
events are to be generated.}

\item{n_origin}{(an integer) Number of origins to simulate.
Default:\code{50}. This is the parameter that has the greatest
influence on the computational time.}

\item{space_resist}{("SpatialPolygonsDataFrame",
"SpatialPolygons", or "sf") Optional features showing
spaces across landscape within which spatial
points (origins) are not allowed. Default: \code{NULL}.}

\item{n_foci}{(an integer) A value indicating the number of
focal points amongst event origins.}

\item{foci_separation}{(an integer) A value between \code{0} and \code{10}
indicating the nearness of focal points from one another. A \code{0}
separation indicates all focal points located in a close proximity
while \code{10} indicates focal points that are evenly distributed across
space.}

\item{p_ratio}{(an integer) The smaller of the
two terms of the Pareto ratio. For example, for a \code{20:80}
ratio, \code{p_ratio} will be \code{20}. Default value is
\code{30}. Valid inputs are \code{10}, \code{20},
\code{30}, \code{40}, and \code{50}. A \code{30:70}, represents
30\% dominant and 70\% non-dominant origins.}

\item{s_threshold}{(numeric) Spatial threshold value. The
(assumed) spatial range within which events are
re-generated (or repeated) by or around the same origin.
Default: \code{250} (in the same linear unit as the \code{poly})}

\item{trend}{(string) Specify the trend direction of
the GTP. Values are: \code{"decreasing"}, \code{"stable"},
and \code{"increasing"}. Default is: \code{"stable"}.}

\item{first_pDate}{(in \code{"yyyy-mm-dd"} format).
Date of the
first seasonal peak of the time series.
Default value is \code{NULL}, in which a
seasonal cycle of 180 days is utilized. That is,
a first seasonal peak of 90 days.}

\item{slope}{(string) Slope GTP trend if
"increasing" or "decreasing" trend is specified.
Values: \code{"gentle"} or \code{"steep"}.
Default value is \code{NULL} (i.e., for \code{stable} trend).}

\item{...}{additional arguments to pass from
\code{gtp} and \code{walker} functions. Arguments from
\code{gtp} can be used to define the nature of the
temporal trend and pattern over time, while arguments
from \code{walker} can be utilized to define the properties
of event generators (walkers) across the landscape.}

\item{show.plot}{(TRUE or FALSE) Whether to display
the plots after execution.}

\item{show.data}{(TRUE or FALSE) To show the output data
Default is \code{FALSE}.}
}
\value{
Returns the global temporal pattern
}
\description{
Generate point pattern in space and time
from scratch based on a specified temporal pattern and
spatial properties.
}
\details{
A list containing \code{k} entries,
where \code{k} is the length of \code{n_events} (see above).
}
\references{
#https://online.stat.psu.edu/stat510/lesson/6/6.1
}
