% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psim_real.R
\name{psim_real}
\alias{psim_real}
\title{Stpp from real (sample) origins}
\usage{
psim_real(n_events, ppt, start_date = NULL, poly = NULL,
s_threshold = NULL, n_origin=50, resistance_feat, field=NA,
p_ratio=20, crsys = "CRS_string")
}
\arguments{
\item{n_events}{(integer) Number of points
(events) to simulate. Default: \code{1000}.
A vector of integer values can be supplied, in the
format \code{c(a1, a2, ....)}, where a1, a2, ...
represent different values.}

\item{ppt}{A 3-column matrix or list containing
\code{x} - eastings, \code{y} - northing, and \code{t} - time of occurrence
(in the format: `yyyy-mm-dd').}

\item{start_date}{Specifies the start date of
the sample data provided (format: \code{yyyy-mm-dd}).
If \code{NULL}, the earliest date
of the \code{t} field of \code{ppt} is utilized.
The end date is automatically set as the 365th day
from the start date.}

\item{poly}{(An sf or S4 object)
Spatial (administrative) boundary covering the area
under study. The default is \code{NULL}, in which an
arbitrary boundary is drawn to cover the spatial extent
of the data. The projection system of \code{poly} is assume
for \code{ppt}, therefore, a user needs to ensure that both
\code{poly} and \code{ppt}(-xy cordinates) are in the same
reference system for accurate result.}

\item{s_threshold}{(numeric) Spatial range
from the origin within
which a walker re-generate events.
Default: \code{NULL}, in which the value is
automatically estimated from the sample data (i.e., \code{ppt}).}

\item{n_origin}{(an integer) Number of
locations to serve as origins for walkers. The value has
largest impacts on the computational time.}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio.
For example, a value of \code{20}
implies a \code{20:80} Pareto ratio.}

\item{crsys}{(string) The EPSG projection code that defines
the xy coordinates (of \code{ppt}). This will be utilized
if \code{poly} argument is \code{NULL}.
See "http://spatialreference.org/" for the list of
EPSG codes for different regions of the world.
As an example, the EPSG code for the British National Grid
projection system is: \code{"EPSG:27700"}.}
}
\description{
Generate spatiotemporal point pattern
from origins sampled based on real dataset.
}
\details{
Returns an object of the class \code{real_spo},
detailing the spatiotemporal properties of a real
sample dataset

#' If not \code{NULL}, a numerical value
based on prior or expert knowledge is expected.
}
\examples{
\dontrun{
data(camden_theft)
#specify the proportion of full data to use
sample_size <- 0.2
set.seed(1000)
dat_sample <- camden_theft[sample(1:nrow(camden_theft),
round((sample_size * nrow(camden_theft)), digits=0),
replace=FALSE),]
#plot(dat_sample$x, dat_sample$y) #preview
result <- psim_real(n_events=2000, ppt=dat_sample,
start_date = NULL, poly = NULL, s_threshold = NULL,
n_origin=50, resistance_feat, field=NA,
p_ratio=20, crsys = "EPSG:27700")
}
}
\references{
Davies, T.M. and Hazelton, M.L. (2010), Adaptive
kernel estimation of spatial relative risk,
Statistics in Medicine, 29(23) 2423-2437.
Terrell, G.R. (1990), The maximal smoothing principle
in density estimation, Journal of the
American Statistical Association, 85, 470-477.
}
