% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stp_learner.R
\name{stp_learner}
\alias{stp_learner}
\title{Learning spatiotemporal properties}
\usage{
stp_learner(ppt, start_date = NULL, poly = NULL,
n_origin=50, p_ratio, crsys = NULL, show.plot = FALSE)
}
\arguments{
\item{ppt}{xyt A 3-column matrix or list of
\code{x} - eastings,
\code{y} - northing, and \code{t} - time of occurrence
(in the format: `yyyy-mm-dd'). Learns the
spatiotemporal
properties of first 1-year subset of the data.}

\item{start_date}{(string) Specify the start date of
the sample data provided (format: \code{yyyy-mm-dd}).
If \code{NULL}, the earliest date
from the \code{ppt} is utilized. The end date is the 365th day
from the specified start date (by default).}

\item{poly}{(An sf or S4 object)
Spatial (administrative) boundary covering the area
under study. The default is \code{NULL}, in which an
arbitrary boundary is drawn based on the spatial extent
of the data.}

\item{n_origin}{(an integer) Value specifying
the number of event origins to synthetize.
Default:\code{50}. Value specified here has the greatest
influence on the computational time.}

\item{p_ratio}{(an integer) The smaller of the
two terms of a Pareto ratio.
For example, a value of \code{20}
implies a \code{20:80} Pareto ratio.}

\item{crsys}{(string) The crs projection to utilize
when \code{poly} argument is \code{not} NULL.
A user can obtain CRS strings for different world regions
from "http://spatialreference.org/".
When both \code{poly} and \code{crsys}
are not NULL, the function utilizes the \code{crs} of the former}

\item{show.plot}{(TRUE or FALSE) Whether to display
the plots after execution.}
}
\description{
Learns specific spatiotemporal
properties of a sample dataset.
}
\details{
Returns an object of the class \code{real_spo},
detailing the spatiotemporal properties of a real
sample dataset
}
\examples{
data(xyt_data)
data(SanF_CRS_string)
#get a sample data
set.seed(1000)
sample_size <- 500
dat_sample <- xyt_data[sample(1:nrow(xyt_data),
sample_size, replace=FALSE),]
stp_learner(dat_sample,
start_date = NULL, poly = NULL, n_origin=50, p_ratio=20,
crsys = SanF_CRS_string,  show.plot = FALSE)
}
\references{
https://www.google.co.uk/
}
