% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stp_learner.R
\name{stp_learner}
\alias{stp_learner}
\title{Learning the spatiao-temporal properties of sample data}
\usage{
stp_learner(ppt, poly = NULL)
}
\arguments{
\item{ppt}{(dataframe)
Spatiotemporal point samples - whose first three attributes
(\code{fields}) are: \code{x} - the eastings, \code{y} - the northing,
and \code{t}, the time of occurence (in the format: `yyyy-mm-dd').
A minimum of (first) three-month worth of data is required. The
longer the data duration, the better the approximations.}

\item{poly}{(as \code{spatialPolygons},
\code{spatialPolygonDataFrames}, or
\verb{simple features}). The boundary (a spatial polygon) surrounding
the extent of the sample points. Both \code{stps} and \code{poly} must be
in the same coordinate reference system. The default is \code{NULL},
meaning that the function creates an arbitrary boundary around
the points.}
}
\description{
Learns all the spatial and temporal
properties (required to build events' population)
from a sample point dataset.
}
\details{
Returns the spatiotemporal details of point datasets
}
\references{
https://www.google.co.uk/
}
