% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker.R
\name{walker}
\alias{walker}
\title{A landscape walker}
\usage{
walker(n = 5, s_threshold = 250, step_length = 20,
poly, coords=c(0,0), show.plot = FALSE)
}
\arguments{
\item{n}{(integer) The number of events
to be generated. Default: \code{5}.}

\item{s_threshold}{(numeric) Spatial threshold value. The
(assumed) spatial range within which events are
re-generated (or repeated) by or around the same origin.
Default: \code{250} (in the same linear unit as the \code{poly})}

\item{step_length}{(numeric) A maximum step taken at a time
by a walker from one state to the next. Should be a fraction
of the spatial units of the landscape. Default: half the size
of the minimum spatial unit in a landscape
(for a constraint landscape) or
\code{Landscape Area/Number of origins * 100} for an unconstrained
landscape. Users are encouraged to input a value that
produce a desirable output.}

\item{poly}{(as \code{spatialPolygons}, \code{spatialPolygonDataFrames}, or
\verb{simple features}). A spatial polygon defining the boundary
within which a walker walks. Default is \code{NULL}, where
a walker navigates an arbitrary square space.}

\item{coords}{a vector of the form c(x, y) giving the
initial coordinates of a walker (e.g. an origin).
Default value is \code{c(0,0)} for an arbitrary square space.
If \code{poly} argument is not NULL, then \code{coords}
assumes the centroid point of the polygon,
unless otherwise specified.}

\item{show.plot}{(TRUE or False) To show the time series
plot. Default is \code{FALSE}.}
}
\value{
Returns a trace of walker's path, and the
corresponding events.
}
\description{
A dynamic object capable of walking
across a constraint or unconstrained landscape,
in line with an in-built transition matrix, and
specified spatial and temporal properties. The transition
matrix defines a two-state status, namely \code{exploration}
and \code{action} states, to indicate the states of
reconnaissance and performance, respectively.
}
\references{
#https://google.co.uk
}
