% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker.R
\name{walker}
\alias{walker}
\title{A landscape walker}
\usage{
walker(n = 5, s_threshold = 250, step_length = 20,
poly = NULL, resistance_feat=NULL, coords=c(0,0),
pt_itx = TRUE, show.plot = FALSE)
}
\arguments{
\item{n}{(integer) Number of events
to be generated by a walker.}

\item{s_threshold}{(numeric) Spatial threshold of
a walker. This is the spatial range within which events are
re-generated (or repeated) by a walker.
Default: \code{250} (in the unit as the polygon shapefile (below))}

\item{step_length}{(numeric) A maximum step taken at a time
by a walker from one state to the next.}

\item{poly}{(An sf or S4 object)
A polygon shapefile within which
event origins are to be generated.}

\item{resistance_feat}{(An S4 object) Optional
shapefile representing spaces across landscape
within which event
origins are not allowed. Default: \code{NULL}.}

\item{coords}{a vector of the form c(x, y) giving the
initial coordinates of a walker (i.e., coordinates
of origins).
Default value is \code{c(0,0)} for an
arbitrary square space.}

\item{pt_itx}{(logical) Check whether any of the
specified initial origin coordinates
falls outside the boundary.
Default: \code{TRUE}.}

\item{show.plot}{(TRUE or False) To show the time series
plot. Default is \code{FALSE}.}
}
\value{
Returns a trace of walker's path, and the
corresponding events.
}
\description{
A dynamic object capable of walking
across any landscape (constraint or unconstraint),
in line with an in-built transition matrix, and
specified spatial and temporal properties. The transition
matrix defines a two-state status, namely; the \code{exploration}
and the \code{action} states. See vignette for details.
}
\references{
#https://google.co.uk
}
